refactor: remove redundant Norm- prefix from components
