+++ frontend/src/components/documents/norms/HeadingFootnotes.vue
