+++ frontend/src/components/documents/norms/VersionList.vue
