+++ frontend/src/components/documents/norms/VersionWarning.vue
